Fix misc. minor sign-comparison warnings

+++ src/crypter.cpp
@@ -31,7 +31,7 @@ bool CCrypter::SetKeyFromPassphrase(const std::string& strKeyData, const std::ve
         i = EVP_BytesToKey(EVP_aes_256_cbc(), EVP_sha512(), &chSalt[0],
                           (unsigned char *)&strKeyData[0], strKeyData.size(), nRounds, chKey, chIV);
 
-    if (i != WALLET_CRYPTO_KEY_SIZE)
+    if (i != (int)WALLET_CRYPTO_KEY_SIZE)
     {
         memset(&chKey, 0, sizeof chKey);
         memset(&chIV, 0, sizeof chIV);
